requirements.txt
streamlit==1.55.0
openai-whisper
pypdf
scikit-learn
torch
